requirements.txt
dlib==19.24.2
numpy==1.24.2
Pillow==9.4.0
pystan==3.7.0
fastapi==0.109.0
uvicorn==0.22.0
prophet==1.1.3
joblib~=1.3.2
pandas==2.0.1
plotly==5.14.1
yfinance==0.2.18
pydantic~=2.5.0
scikit-learn~=1.3.2


